Refactor the rendering pipeline (again)
Introduce trait Primitive, implemented by geometric primitives
(currently Face, LineSeg, and Sprite). Add struct Batch to collect
a list of primitives and associated vertices to be rendered.

This change deduplicates and centralizes the rendering logic and
makes the high-level pipeline easier to follow. It also enables
rendering several objects in one batch, or a single object in
several batches. This is useful for depth-sorting transparent
objects, among other things.