Replace VectorLike with traits Affine and Linear (#1)
The `Affine` trait represents elements of an affine space: a generalization
of vector space where there is no privileged origin or zero vector.
Subtracting two affine values returns a separate difference type;
a difference can be added to or subtracted from an affine value, but two
affine values cannot be added together.

The `Linear` trait represents vectors, elements of a vector (linear) space.
`Linear` is a subtrait of `Affine` and adds the requirement to have a zero
vector and additive inverse (negation).

The difference between `Linear` and `Affine` is that between vectors that
represent translations (directions with length) and vectors representing points.
Adding two points is not meaningful, but subtracting two points yields
a translation, and adding a translation to a point yields another point.

In the common homogenous coordinate representation of vectors, points have
the w component equal to 1 and translations have it equal to 0. Thus, translations
are not affected by translation (an affine operation), but points are.