Add a generic color type
`Color` is an `Affine` type very similar to `Vector`; it is generic over
representation and color space or basis. SRGB(A) colors with u8 or f32
components are currently supported, as well as HSL(A) colors, but no
conversions yet. u8-based colors are not linear, only affine, as u8
cannot represent signed  differences. i16 is used as the difference type.
Color components are also bounded: they require clamping before displaying.

Todo:
* Proper `Affine` impl and varying support for integer-based colors
* Conversions between RGB(A) and HSL(A)
* Spaces and conversions for linear RGB(A) and pre-multiplied alpha
* More support for different output pixel formats (u32/[u8] and BGRA etc.)
* Colors with 16-bit (u16) channels and palette-based colors?