Add a 2D buffer and slice types
A Buf2 is backed by a Vec and owns its contents, while Slice2 is its
counterpart borrowed type, referring to a rectangular region in a Buf2
or any suitable block of memory. Slice2Mut is the mutably, exclusively
borrowed version of Slice2.

Notionally, a Slice2 is a reference to a list of (1D) slices of the
same length that need not be contiguous in memory. The spacing between
the starts of two consecutive slices in a Slice2 is called its stride.

From the perspective of Rust, Slice2 borrows a single contiguous slice,
which means the borrows of two 2D slices may overlap even though their
rectangular 2D areas do not. As usual, neither of the borrows can be
exclusive in that case.