Implement triangle rasterization
Rasterization is the process of turning a vector shape into discrete
pixels on a framebuffer. Rasterization proceeds by turning a primitive
such as a triagle into a sequence of *scanlines*, each corresponding
to a horizontal span of pixels covered by the primitive on a given line.
The scanlines, in turn, are converted into a series of *fragments* that
represent potentially drawn pixels.

If depth testing (z-buffering) is enabled, the fragments are then tested
against the current depth value in their position. For each fragment that
passes the depth test, a color is computed by the fragment shader and
written into the framebuffer. Fragments that fail the test are discarded.