Add traits for vertex and fragment shaders (#17)
Shaders are functions that are used to customize vertex and fragment
handling during rendering.

A vertex shader is responsible for transforming and projecting each
vertex in the rendered geometry, usually using a modelview matrix to move
the vertex to the view (camera) space and a projection matrix to transform
it to the clip space. Vertex shaders can also perform any other per-vertex
calculations and pass on the results as attributes of the output vertex.

A fragment shader is used to compute the color of each individual pixel,
or fragment, drawn to the render target. The fragment shader receives as
input any vertex attributes interpolated across the primitive being
rasterized, such as color, texture coordinate, or normal vector.