Add cubic Bezier curves and splines
A Bezier curve is a parametric curve defined by a number of control points.
The first and last points are the endpoints of the curve, while the middle
points control the shape of the curve by "dragging" it towards them.

This implements cubic Bezier curves, defined by four control points.
Features include:

* The control points can be of any Affine type
* Forming a spline by joining several curves
* Computing the tangent of the curve
* Adaptive approximation of a curve with a polyline