Add floating-point function adapter
The functions exposed in core::math::float::f32 automatically delegate to either
std, libm, or micromath, depending on what features are enabled. This lets code
be agnostic of which implementation of fp functions is actually used.

There is also a fallback implementing a few common and simple functions even
when none of the fp features is enabled.